+++ NEWS.md
@@ -1,6 +1,6 @@
-# jsmodule 1.4.5
+# jsmodule 1.5.0
 
-## New module (Thanks for [Yoonkyoung Jeon](https://github.com/cyk0315))
+## New module: Subgroup analysis with forestplot (Thanks for [Yoonkyoung Jeon](https://github.com/cyk0315))
 
 * forestcox (`forestcoxUI`, `forestcoxServer`)
 * forestglm (`forestglmUI`, `forestglmServer`)